Added the dDockTabs class. This is analogous to the dDockForm, in that it is based on the wxPython AUI; in this case, the aui.AuiNotebook class.
So far this has only been tested on OS X and wxPython 2.8.4.0; I plan on testing it on other platforms soon. I haven't exposed any of the tab styles, and it also appears that the tabs can only be placed on the top. But they are fully movable and dockable.