Began implementing toolbar support for dDockForm, but for the life of me I can't see a difference between toolbars and any other panes. Added the 'Toolbar' property that returns True if the pane is a toolbar.
Added the '_Manager' property to _dDockPanel. Previously the only reference to the AUI manager available was through the form.

Fixed several naming conflict issues. Besides the wxPython requirement of siblings having unique names, the AUI stuff requires that each managed pane be uniquely named (they are actually children of the floating pane window).

Fixed some issues with some properties not being set; adding them to the delayed properties resolved that.

Added support for saving and restoring the layout of a dDockForm when SaveRestorePosition is True. This is still not working as smoothly as possible; if you change the properties passed to a _dDockPanel, restoring the layout also restores the old property settings.

Tried to implement DynamicCaption as a proof of concept for Dynamic* props, but there is no way to tie into the AUI events such as docking and undocking, so this is marginal at best right now.

Added the 'ShowActivePanel' and 'TransparentDrag' properties to dDockForm. When the former is True, the 'active' panel's title bar will be hilited. When TransparentDrag is True, the panel being dragged assumes a semi-transparent appearance.

Fixed the form's update(), as it wasn't propagating to the panels.