Removed the ReleaseOnEscape property from dDialog. Added the CancelOnEscape property to dStandardButtonDialog; when this is True (default), pressing the Escape key will be handled by the appropriate standard button handler.
Re-instituted the ButtonSizerPosition property.

Deprecated onOK() and onCancel(). The methods that will be used by developers to react to standard buttons being clicked are now named run*; e.g., runOK(), runHelp().

Fixed a bug that has been around in dDialog for years. It used to be that pressing the Escape key destroyed the dialog if ReleaseOnEscape was True; the code in the class would still try to access the now-destroyed object, throwing an error. Now dialogs are never destroyed in these classes; if you create them, you have to manually call their release() method.