Removed the raising of the ValueError if you set a menu item Caption in the form of "Open\tCtrl+O".
Instead, simply override any existing HotKey with the hotkey provided in 
the string, and only set the caption part to the Caption property.

The only potential problem with this is in this example:

item.Caption = "Open\tCtrl+O"
...
item.Caption = "Close"

In this case, the item will still have the Ctrl+O hotkey, and if you 
didn't want a hotkey, you'd have to then set it explicitly with:

item.HotKey = None

I think allowing the shortcut \t format doesn't harm anything, so why not
leave it in. If there's serious disagreement to this, let's give a 
DeprecationWarning instead of the ValueError, as the \t method was the
only way to set the hotkeys until relatively recently and I know all my
apps use it still - I'm sure there are other apps out there as well.

Paul