Added back a try: block that was removed in [4184] to stuff the getBlankValue() into self.Value if calling the instance method raises dException.NoRecordsException.
This fixes one of my problems with HEAD. I have a bunch of functions
in the UI that are thin wrappers to methods of the bizobj, and 
these ui functions will run with the ui update() calls, but I 
can't seem to control when update() will happen, and it sometimes
seems to happen when the dialog is hidden, or before the dialog is
ready for display, or after it is already being destroyed or hidden. 

I don't want to catch biz.NoRecordsException in each one of these ui 
wrappers; I just want the controls to get the blank value for now,
as it isn't an actual problem (after the dialog is shown, the controls
update themselves with valid values from the bizobj functions).

Also did some minor housekeeping around that section of the code.