For some reason, the dShell form was intercepting key characters with "Ctrl+Character" and stopping the KeyDown event from firing on Windows.  Now, instead of catching the KeyDown event and looking for the specific sequence, the form's bindKey method is used instead.  This works on Windows.  I have bound cmd+R to Mac, but I need someone to test the changes.
Also, I fixed a typo in _LookupPanel's onListKey.  dKeys.arrows should be dKeys.arrowKeys.  arrowKeys is a dict, so we need to reference values. Again, test the changes on Mac.

Lastly, on Windows, I am still having trouble in onListKey with the return and numpad_enter key events propagating.  Events are fired as getEventData is called, but something is stopping the event before it makes it through.