I found the filt argument to setAll() to be hard to use: the expression has to be written so that the leftmost word is a member of the current child object being evaluated, and any names in the expression need to be local to dPemMixin, not the caller. So the simple case in the docstring works, but how to evaluate if the child is an instance of my special subclass?
I couldn't do filt="'TxtFraction' in str(BaseClass)" because of the 
requirement that the left-most word must be an attribute or member of the
child being evaluated.

I couldn't do filt="BaseClass == TxtFraction" because TxtFraction
doesn't exist in dPemMixin's namespace.

So, I added a instancesOf argument, that accepts a single or sequence
of class(es), and added logic to setAll() to handle this. It shouldn't
have any effect on current uses of filt, but I feel this better satisfies
the common case of wanting to call setAll() only on instances of 
specific classes.

Perhaps an improvement to filt would be to add a placeholder, instead
of it being fixed to the left. Another improvement could be to pass in 
the locals and/or globals to setAll(), so the eval() takes place in 
the environment of the caller.