Made it simple to create custom grid editors using Dabo controls.
To create the editor, just pass the control that you want to use to dabo.ui.makeGridEditor(). It will return a class that you can then assign to a column's CustomEditorClass property.

Since editors by default will be sized to the grid cell they are editing, I added two optional parameters to makeGridEditor(). You can pass 'minWidth' and 'minHeight' values; the class will then ensure that the editor is displayed with at least those dimensions.

Renamed the dFoldPanelBar/dFoldPanel classes to dSlidePanelControl/dSlidePanel, repectively. Added deprecation warnings if the old names are used.

Lots and lots of improvements to these classes; too many to list. I'm still not 100% satisfied with them, though, and will continue to work on them, but they are stable and work as they should.