Fixed a long-standing bug in this composite control. There were several situations where it appeared that the spin buttons stopped responding to clicks, even though the displayed value was within range. The problem turns out to be that while we were correctly setting the textbox's value, we were not setting the spinner's value; instead, it was incrementing by +1/-1 for each click. This commit fixes that problem by eliminating the spinner's settings from the mix; instead, we govern min/max/wrap through our own code.
I also fixed the problem with the collision between float and Decimal values. All values in the comparisons are now cast to the type of the edited value when there is a conflict.

Removed the conditional import on the decimal module, as we are now requiring Python 2.4.

Fixed the missing docstring for the class, and added some other docstrings to methods.

Renamed dSpinButton to _dSpinButton, as it is not meant to be a public class, and is not available in dabo.ui.uiwx.__init__.py.