Fixed Trac Issue #1165: opening a saved form does not correctly restore the size of that form.
Fixed a bug in the ClassDesignerEditor that made it hard to delete a method just by deleting the body of the method.

Fixed the cause of most of the sources that caused the message: "None PROP DICT:, ROW=..." to print out while starting the Class Designer. It was mostly the result of an overzealous wx.Grid trying to determine CanGetValueAs() when the grid was empty. IOW, not an error condition, and thus doesn't merit the output. There are still some timing issues when creating sizers; I've suppressed that output under the verboseLogging setting because it is simpler than getting into the underpinnings of sizer creation.

Added some additional support for dSlidePanelControl. It still is not 100% solid, but at least it's somewhat workable for now.