Improved VirtualFields by adding the option to set a given VirtualField to a dict instead of a function. Currently, this dict has one required key, "func", that stores the function object, and an optional key, "requery_children", that determines whether or not child records are requeried before the virtual field function is run.
Changed default behavior to *not run* the child requeries, as this 
was a performance hog for an uncommon case. If your VirtualFields
function requires that child records have been requeried before 
running, then you can either requery them inside the function itself,
or define the VirtualFields like the following:

old:
  self.VirtualFields["myfield"] = myFunc

new:
  self.VirtualFields["myfield"] = {"func": myFunc, "requery_children": True}

TODO: Note this in the docstring.