Added the deleteByValue() method to dPref. This enables you to delete all pref entries that contain the passed value.
Updated dShell so that if you re-use a command from its history, the old entry for the command is deleted. The effect is that the history only contains one instance of each command, with the most recent usage at the top.

Fixed a bare-exception removal bug in dGrid.buildFromDataSet() that didn't catch the case where the keyCaption dict was None.