Added 2 new properties to the dEditor widget.  They are ShowEdgeGuide and EdgeGuideColumn.  Basically, ShowEdgeGuide will display a line at the specified column position.  Note that if the editor has a variable spaced font set, wx uses the average of the characters' widths to determine where to put the line.  Also note that wx will only allow for either HiliteCharsBeyondLimit or ShowEdgeGuide to be True, not both.  So, when one is set to True, the other is set to False.
Fixed a bug where if we set HiliteCharsBeyondLimit to True and then set it to False, the highlighting wouldn't go away.

What I would like to do is combine HiliteLimitColumn and EdgeGuideColumn into one property.  I would also like to combine HiliteCharsBeyondLimit and ShowEdgeGuide in a property (EdgeMode maybe?).  Ed and Paul, thoughts?  If y'all agree, I'll do it, and keep HiliteCharsBeyondLimit and HiliteLimitColumn functional, but issue a deprecation warning when they're called.