This fixes the updateBox traceback, but I left it commented as I don't know if there was some special reason we were calling _paintHeader() explicitly from the ClassDesignerControlMixin.
Now that this is fixed, I'm getting another traceback in the class designer
that I'll pass back to Ed:

{{{
Traceback (most recent call last):
  File "/usr/lib/python2.5/site-packages/wx-2.8-gtk2-unicode/wx/_core.py", line 14551, in <lambda>
    lambda event: event.callable(*event.args, **event.kw) )
  File "ClassDesigner.py", line 2814, in addNewControl
    obj.layout()
AttributeError: 'controlMix' object has no attribute 'layout'
}}}

This is after right-clicking on the form, choosing "add grid" from the Data Controls submenu,
and accepting the default 3 columns. The grid appears but this traceback appears in the
terminal.