Ok, this fixes the error I was getting on deletion of the editor object in relation to the bookmarks.  I eliminated the timer.  We didn't need it.  In addition, I eliminated the bookmark saves on deletion and on saving the file.  The were irrelevant and unnecessary considering we save the bookmarks after every bookmark operation (clear, clear all, and set).
This solved the error, but I was still having some issues with the saving bookmarks.  No matter what I did, I could only get the UserSettingProvider to save one bookmark key.  Not sure of the problem, but the solution was to save the bookmarks as a list of tuples.  It works out better anyway because there is less database reading and writing and allowed me to eliminate some clunky code.

Tested the IDE tools with the change and they work as expected.