Made some fairly hefty changes to the way that bookmarks are shown.  Before, they used to appear in the margin used for line numbers.  However, if you had a document that went into the 1000's of lines, setting a bookmark on line 1000+ would hide the line numbers.  Also, if you set UseBookmarks to False, you couldn't hide the bookmark icons.  Furthermore, you would hide the bookmark icons if you set ShowLineNumbers to False.
I have created a new margin for the bookmarks to be shown in.  This margin will be visible/not visible based on what UseBookmarks is.  I also capture left clicks in the margin.  If the line number margin is click, it behaves as before.  If the bookmark margin is clicked, it fires a stub called processBookmarkClick and passes the line number and the existing bookmark name (if no bookmark, it is passed None).  The user can decide what to do in the subclass.

I also added 3 new properties to the editor.  The first is BookmarkIcon.  You can set the symbol used as the icon.  BookmarkBackColor and BookmarkForeColor are what you expect them to be and take a valid string or rgb tuple input.  The properties can be set in the constructor or anytime afterward.  Defaults are still the ugly cyan circle with the grey border.