These are the changes to the framework required to create web-enabled applications in Dabo.
There should be no effect on any existing application, but please test thoroughly, and report any issues. I have tested every app of mine over the past several weeks without incident.

Web-enabled apps will require that the simplejson module for Python be installed. If you do not create a web app, you do not need to install this module.

The server side of this is just about ready, and will be committed soon. The tools for creation of web apps are still on the drawing board.