Changed the default for SQLite to not create the database file if it doesn't already exist. This setting is controlled by dabo.createDbFiles, which defaults to False.
Added dException.DBFileDoesNotExistException class. This is raised if you try to connect to an SQLite file that does not exist and dabo.createDbFiles is False. This exception can be caught in the calling routines and handled appropriately.

dConnection now takes an optional constructor parameter 'forceCreate'; if that is True (default=False), the db file will be created no matter what the system setting is. 

dbSQLite.getConnection() takes that same parameter, and passes it to the connection constructor.

dPref, which is supposed to create the DaboPreferences.db file no matter what, has been modified to pass  forceCreate=True to the connection.