Added the 'cloneRecord()' method to dBizobj. This will make a copy of the current record and append it as a new record in the dataset. The copy will contain all of the columns of the current record except for PK columns.
Added the 'appendDataSet()' method to dCursorMixin. This assumes that the dataset to be appended has the same table structure as the current dataset; if not, an error is raised.

Added the 'isRemote()' method to dBizobj. This returns True if the connection for the bizobj is a remote web connection.

Implemented remote proxy support in the getFieldNames() method.