Cleaned up a lot more unicode issues, especially the one reported by Alexander Altanis when he tried to use Dabo with Greek text.
Also fixed a bug when a field in the DataSource is named the same as a method in the DataSource. Previously we looked for methods in the DataSource first, and then tried to call get/setFieldVal(); this would cause Dabo to call the method instead of getting the field value. Now we look for fields first, and only look for methods if there is no such field in the dataset.