Added the ability to filter data in the current dataset without requerying the database. In order to ensure that the UI is properly updated, you should call dForm.filter().
There are still issues to resolve when using filter(). One situation is if you change a record, and then apply a filter that 'removes' that record from the current data set. Most of our scan routines are based on RowNumber, and this is completely broken by filtering. So for now I'm adding this with the warning that some things affected by this are not rock-solid by any means.