Reworked the restoration of form size, position, and state, inspired by the patch submitted by Jacek Kałucki.
We don't need to wait until the form is activated anymore, as Name will have
been set somewhere in the initProperties() cycle. This change results in a
much smoother form instantiation, as the form isn't shown first in the 
default position, only to jump to the saved position.

I wrote this change on Mac, where it seems to be working sanely. I'll now
test on Linux and Windows.