Made some usability improvements to ReportDesigner:
1) If an object is selected, and <enter> pressed, focus to prop sheet 'expr' prop.
   (previously, <ctrl>+<enter> would do this). Also, double-click still does this.

2) If the propsheet 'expr' property was activated as in #1, once the user is done
   editing the property, focus automatically goes back to the designer.

These changes enable working with the report designer without needing the mouse or
a lot of extra key combinations. Note that <tab> and <shift>+<tab> will iterate
forward/backward through the objects in the designer, so my workflow today has been:

1) Add a string object, copy/paste 8 times.
2) Select top object, press <enter>
3) Edit expr property, press <enter>
4) Press <tab>
5) Repeat steps 3-4 8 times.

I'm working on a report with over 120 individual objects in the detail band, with
stacks of 8-9 like items. So working on this caused me to wish for better usability,
hence this commit.