Added new Report properties "Author", "Subject", and "Keywords". Enhanced the PDF output to push the value of these properties, plus "Title", to the PDF which results in the title being displayed in the preview window, and the rest of the properties being accessible from the properties window of the pdf viewer.
I finally got tired of seeing "untitled" in the title bar of my pdf viewer...

Added SpanningLine report object. This can go in the pageHeader or any
group header. You specify the x,y relative to the header band as usual,
and a second coordinate to draw the line to in the corresponding footer
band (xFooter, yFooter).

Currently SpanningLine isn't exposed properly in the ReportDesigner, but
if you add it manually to the rfxml you can then select it in the object
tree and edit its properties in the property window of the designer.

Here's an example from my report:

 98           <SpanningLine>
 99             <x>0</x>
100             <xFooter>0</xFooter>
101             <y>9</y>
102             <yFooter>2.5</yFooter>
103           </SpanningLine>
 
This appears in the groupHeader band for one of my groups, and results in
a vertical line being drawn from (0,9) in the header to (0, 2.5) in the 
footer. It does this across pages, too. If we change <xFooter> to, say, 10,
the line will no longer be vertical but will slant from top to bottom.

I'll add a SpanningRectangle in the future, if there's demand for it.