Put beforeInit, afterInit, and initEvents into the private _beforeInit, etc. methods, to allow for appdev's to add custom code to these hooks without accidentally overwriting base funcionality.
Defaulting the value to today's date in afterInit() was reportedly overwriting
the value in the database. Moved this to _beforeInit() and default of None,
which I think is a better default than assuming today.

Took away the if statement that only acted on the shortcut keys if there was
currently a GetValue() greater than u'', because we'd like it to try to work
even if there isn't yet a value.

I don't use dDateTextBox, although I just switched a couple apps over to use it
instead of the raw dTextBox, so I think this commit works well but please test
with your own apps and let me know.