Added MajorProperty property to report objects, which tells the ReportDesigner which property to edit when double-clicked.
Refactored some things in ReportDesigner, moving the editProperty method out
of the designer form and into the controller (app).

Now, you can double-click on objects in the tree, as well as on the design 
surface, and the property editor will gain focus to whatever property is
set as the MajorProperty (sensible things like expr for string objects).