Added some events to dReportWriter: ReportBegin, ReportEnd, and ReportIteration. Your code can bind to them like any other Dabo event.
Added PrintStatus property, which if True (default is False) will
print status information to stdout like "processing record 1 of 20".

Added ShowProgress property, which if True (the default) will
show a progress gauge during the processing of the report, and hide
it when done.

There's a cancel button on the progress window, but it doesn't work.
I think I need to put the processing of the report into a worker 
thread for the cancel button to be responsive. I'll try some more
later.