*MAJOR* change to the underlying cursor engine. This will need to be tested!
I looked into the problems with binary values not being able to be saved, and was surprised to find that internally, we are still doing string substitution to build our SQL statements to be executed. I thought that Carl Karsten had re-written all of this years ago, but apparently not.

dCursorMixin now passes field values as parameters to the dbapi's execute, instead of trying to create a fully-substituted string that incorporates the field values. This was the primary reason that Dabo was not able to handle binary values correctly.

I've added a backend-customizable attribute called 'paramPlaceholder', which is the character to be used for parameterized execution. The default in dBackend is '%s', and I've overridden it in dbSQLite to be '?'. Please update any other backends that don't use '%s' for parameters.