Fixed the issues with custom classes in the Class Designer not being properly inherited when running inside other class designs. The problem came down to pathing, as have most similar issues, so I approached it by adding some smarter pathing code.
There is now a method in lib.utils named 'locateRelativeTo()' that will take a main file and a target file, and try to find the target file, using the main file's location as a starting point. It will also look in the standard subdirectories of a typical Dabo application for the target file. If found, it returns the full path to the target. If not, it returns the original target file path.

I also consolidated the logic for app standard directories. Recent changes had not kept the different locations where they were referenced in sync, so now there is an attribute of dApp called '_standardDirs' that is a tuple of the subdirectory names. There is also now a method of dApp called 'getStandardDirectories()' that will return a tuple containing the HomeDirectory and the full paths to all these subdirectories.