At one time in Dabo's life, having the SQL as part of the error message displayed to the user may have made sense, but now I'm running into real trouble with it:
My SQL statements can be huge (100+ lines) and on at least one platform (Mac) the
dialog displayed gets sized below the screen, and the Okay button can't be exposed.

Also, we now have dbActivityLog, where we can log such errors along with the SQL. 
In general, the user that will see the message only cares to see the actual error
message (maybe not even that) but will be confused with all the noise of the SQL
statement.

So, I've removed the SQL from the dbQueryException raised from execute(), and 
added the SQL and error to the dbActivityLog.

We should also look into the dialog size problem on at least Mac, as a separate
issue.