Several major changes in this commit. Please test, and also note that I have once again made changes that could break your .cnxml files, due to a change in the way that encryption is handled.
The other big change is in the handling of pathing. I'm focusing my efforts on making the concept of an app's HomeDirectory much more important as the starting point for all path-related matters. So again, any relative pathing you may have in your code, especially in .cnxml and .cdxml files, may need to be updated to bring those paths in line with the current approach.

Recent changes to the SimpleCrypt script allowed you to set a dabo-level attribute in your settings_override script to your encryption key, and if that was available and you had the PyCrypto package installed, it would use DES-level encryption in your app.

This bothered me on several levels, but mostly because it required a plain text file that contained the key. In other words, while the encryption was better, it was no more secure.

Now there is a write-only property of dApp called 'CryptoKey': when you set that property, if you have PyCrypto installed, it will create a new SimpleCrypt instance that uses DES3 encryption. And to avoid having to pass a plain-text value to the app creation, you can set this property at any time, passing either a string or any callable. By passing a callable, you can more effectively hide the way your key is stored; the actual key never has to appear in your code at all.

All of the util functions to create and/or resolve relative pathing have been updated to default to the app's HomeDirectory instead of the current working directory. Also, many of the routines that call those methods have been udpated to work with the new relative pathing scheme.