I just fixed a long-standing bug witn grids on Windows. Could be that this fixes #1314 too, but haven't tested that. That ticket also deals with onSelectCell firing, changing the row number to 0.
Seems that sometimes, the incremental search will get set to "\r" after the 
<enter> key is pressed. The search timer will fire, and then "\r" will be
searched for, changing the row number of the bizobj.

In my case, it was nasty: <enter> would go into edit mode on the current row,
displaying my modal edit dialog. But the search timer would still fire even 
though the modal dialog was frontmost, which would run the bogus incremental
search, changing the row number of the underlying bizobj. I was unknowingly
editing fields in a different record than I thought I was!!

This fix probably isn't the best (should have dug into the onKeyChar handler 
that starts the timer in the first place) but it does solve the problem.