Updated the logging changes to not attempt to create any file-based logs. This should fix some of the problems reported on Windows.
Changed the dbActivityLog to mirror the mainLog: it has provisions for both a console and a file-based handler. This will allow you to better control db activity logging.

Renamed the log handlers to include 'Handler' in their name. I think that it is much more clear that 'consoleLogHandler' is a handler; the old name of 'consoleLog' would lead one to believe it is the log itself.