Added a code editor to the dShell class. This is a first pass, but I have been using it in my personal work for several weeks, and it's finally polished enough to commit.
Frequently when working in dShell, I need to run more than a single line of code, which, while possible, is not very easy to do in the shell. Editing across multiple lines is difficult, and you can only run one block of code at a time.

Now you can switch to the code editing page, write and edit the code you want to run, then press Ctrl-Enter (or click the 'Run' button) to push that code into the shell and execute it as if you had typed it there. The code editor maintains a history of code you have executed, which you can access via Ctrl-Up/Ctrl-Down.

I also improved some of the code in dEditor, and added some additional handling so that the code editor in dShell can pick up on the local namespace of the shell.