Ever since introducing the report progress bar, I've received about 3 error logs from customers per month with the 'wx.Yield called recursively' error, and I've never been able to reproduce it on my end.
I did some Googling, and someone in May found a workaround that solved the
problem for him. I've added a _safe=False argument to dabo.ui.yieldUI: when
True, the new code is run. I changed the yieldUI() calls from dReportWriter
to yieldUI(_safe=True) to test it in the field. It doesn't seem to break
anything. I wanted to keep the argument "private" for now until it has a
track record, at which point it'll probably become the default code for
yieldUI().

Here's the link to the article I got the code from:
http://www.python-forum.org/pythonforum/viewtopic.php?f=4&t=18801