Changed some unicode() calls to ustr() calls where it seemed appropriate. This includes calls with no explicit encoding that could potentially try to decode non-ASCII chars, as well as other places where it probably doesn't matter but resulted in simpler code.
Added the exception type to the "Error fetching records" error log entry.
I'm wondering if we shouldn't be eating these exceptions. In my case,
my customers will never see the error log entry, only that no records
were selected. If the exception weren't eaten, I'd be emailed a report
of the exception.

Annotated dabo.lib.ustr() with comments, mostly because I had to spend time
figuring out the rationale for why it is the way it is, and didn't want to
have to do that again.

Made the exception catching in ustr() more narrow, so that non-expected
exceptions will still be raised normally.

Replaced the code in dApp.str2Unicode() to simply call dabo.lib.ustr()
instead.