I'm starting to experiment with biz.cacheExpired(). I'm overriding that method instead of using ChildCacheInterval, because I have some rules in place to determine if the cache is expired or not, that actually don't involve amount of time since last requery.
Sometimes in my code I explicitly call biz.requeryAllChildren(), and the 
children weren't requerying because of code in my overridden cacheExpired().

So, I added a force arg to requeryAllChildren(), default False, so that when
my code explicitly calls requeryAllChildren(), it can send force=True to make
sure the children get requeried.

I also annotated a setCurrentParent() call that I believe should happen
unconditionally but I didn't want to make the change without Ed's blessing
because there could be a good reason for it being the way it is that I just
don't see at the moment.