I added a parameter to the _clearDocument method of the dEditor object that specifies whether or not to clear the undo buffer.
I got really annoyed because saving a file would clear the undo filter.  If I made a change, saved the file, tested the change and it didn't work, I would want to sometimes undo the change.  That couldn't happen if the undo buffer was cleared on save.  So, we specifically don't clear the undo buffer on save, but still clear it everywhere else.