Removed the recently-added force argument from requeryAllChildren, in favor of adding a note to the docstring that developers should explicitly call self.expireCache() before requerying, which ensures all cursors (not just the current one) in a bizobj will get requeried next time the record becomes active.
Moved the setCurrentParent() call outside the check for RequeryWithParent,
as the setting of the parent should happen unconditionally. However, 
this will result in the side-effect of an implicit requery() if the 
cursor for that key doesn't exist yet, *and* RequeryOnLoad is True.

Refactored cacheExpired() a bit, with a check for lastRequeryTime being
None.

Added biz.expireCache() to dForm.requery() per Jacek's suggestion.