Added the VerticalHeaders property. When True, the header captions are written vertically; i.e., at a 90 degree angle. This can be useful if you have narrow columns and need to display a label for the column. Thanks to Bronwyn Woods for the idea.
Added the AutoAdjustHeaderHeight property. When True, the HeaderHeight of the grid will change to fit the Captions, whether they are written vertically or not.

Added the SortIndicatorSize and SortIndicatorColor properties, which allow you some control over the appearance of the little sort triangle displayed in the header when that column is sorted.