Improved the handling of the WordWrap property in dEditBox. Fixed some punctuation in the Love Boat lyrics in the test code that had been annoying me for some time. NOTE: the horizontal scrollbar doesn't seem to be drawn in OS X.
Fixed test.py to set the width of the frame, and not the object themselves. Previously, the MinSize for the object was being set via an explicit Width setting, making resizing problematic. Now the form is sized, allowing the sizer to handle the sizing of the object.

Added some saner minimum sizes in dPemMixin. Also eliminated an unnecessary call to get the color tuple for 'black'.