First pass at creating an Object Inspector, which is largely modeled on the wxPython Widget Inspection Tool (http://wiki.wxpython.org/Widget%20Inspection%20Tool). It largely works as expected, but I'm sure there will be cases that need cleaning up.
To invoke it, I've added a menu item to dBaseMenuBar that can be triggered by Ctrl-Shift-I.

Right now the form sections are arranged using splitters, because the form is cdxml-based. This allowed me to rapidly prototype it, but I would prefer to use dDockForm instead, which unfortunately is not yet supported in the Class Designer.

Feedback welcome!!