Enhanced the ReportDesigner by making Shift+Enter on an object activate the propsheet editor for the currently selected property in the propsheet.
(Plain Enter has always activated the propsheet editor for the object's
MajorProperty - this makes it so I can edit a bunch of x values, for example,
without the propsheet jumping off the property every time.)

This would work for shift+double-click if the shift+click hadn't already 
deselected the object. We need to somehow fix the double-click event to have
a chance to fire *before* deciding it is just a single click, but that is
at the Dabo level.