Deprecated Frameset/Paragraph; Added Memo which accomplishes the same thing in the same way but the appdev doesn't need to know about the hierarchy.
Removed the auto-updating of the _def default properties when the main
property changes, as it was causing unneeded cruft in the rfxml.

Changed default expr of String and Memo to "String" and "Memo", 
respectively.

I'm going to think about whether or not to raise deprecation warnings
upon encountering Frameset/Paragraphs. I may just offer the conversion
in the designer and leave Frameset/Paragraph in the report writer 
forever for backward compatibility.