I spent literally all afternoon chasing down a segmentation fault/illegal instruction in my code, and it turns out a copy/paste error added the same sizer twice:
187     vs.append(hs, "expand")
188     
189     hs.appendSpacer((2,0))
190     hs.append(dabo.ui.dCheckBox(self, ##DataSource="ord_misc", ##DataField="enabled",
191         Name="chkEnabled", Caption="Enabled"), "expand")
192     hs.appendSpacer((20,0))
193     vs.append(hs, "expand")

Notice that in 187 I append hs, and I again append it in 193. Obviously, I had 
meant to create a new instance of hs in 188, but I didn't notice because the form
displayed okay. It was only when closing the application that the segfault
would occur. 

So, given that a simple check in the sizer insert() code would have saved me all
this grief, I offer these two lines to my future self and hopefully others, too.