Added new property RequeryChildrenOnNavigate to dBizobj, allowing a way to (temporarily) suspend requerying any child bizobjs when navigating the bizobj. Of course if an appdev does this, they are responsible for turning it back on and/or calling requeryAllChildren() explicitly.
Reworked the navigation data update code in dForm to take advantage of
this property to navigate the primary bizobj quickly. Now, while the user
is navigating, the frm.afterPointerMove() will fire as the navigation
occurs, allowing appdev code to respond with a quick UI update to that
effect. But only after the DataUpdateDelay has passed will the children
be requeried, the RowNumChanged event be raised, and the form updated.

Reworked some code in datanav and added properties allowing the automatic
setting/unsetting of the new bizobj property depending on whether or not
the browse page is active, allowing for quick scrolling of the main grid.

I've tested this on Linux with my app and with a DataUpdateDelay of 250
it is really a very nice improvement. But of course we need to test on
other platforms as well.