Resolved an inconsistency that's been bugging me for awhile. Dabo was defaulting SelectOnEntry to False for text controls, but not really.
Windows and Linux were still selecting all text when the control
was navigated to, but Mac was not.

This commit defaults the property to True, so that at least by 
default the property behaves as it should. We may need to look
further into this to make setting it to False work correctly on
Windows and Linux though, since apparently there's something in
the underlying platform toolkits that perform this automatically.