Fixed a problem due to an obscure inconsistency in the wx class.
There are 6 settings that affect link appearance: the color and the underline status of a link, in each of three states: link, visited, and hover. These are managed at the wx level through GetColours(), SetColours(), GetUnderlines(), and SetUnderlines(). For each you either pass in or are returned a 3-tuple of values. In all the order of the values is (link, visited, hover) - except for GetUnderlines, which returns (link, hover, visited). I've fixed the code to account for this odd inconsistency.

I also modifed the setting code for these values to do nothing if the new value is the same as the current state.