This is an overhaul of the MenuDesigner that I've been working on for a while. It's finally at the point where it pretty much works as it should.
I cheated a little from my original idea, which was to create panels with the menu item captions and hot keys drawn on them, but that proved to be too slow to get working. So I used hyperlinks, taking advantage of their clickability, even though their appearance is not all that controllable. I plan on replacing the hyperlink-based classes with something that looks much better, but for now this is working.