Optimized the requery() cycle to not correct the field types immediately. Instead, the field types are corrected as needed during the getFieldVal() cycle. In my testing this yields a 7x speedup when requerying 1000 rows in a table with 30 fields. We pay for it later a little bit when the field type gets corrected and a flag set during getFieldVal(), but the key is we don't pay for it all at once, and we don't pay for something we may never use.
Optimized some use-cases of getDataSet(). Previously, we used a subtractive 
approach by copying each record dict and deleting the fields we weren't
interested in. Now, we use an additive approach which results in a 3.25x
speedup when calling getDataSet() on 500 rows and only asking for 2 fields
out of 30. Other use cases are comparable to the old speed - a little 
slower since we need to correct the field types if needed during the cycle
of desired rows.

Tested on MySQL and SQLite. Please let me know if you notice any trouble.